Lançamento da primeira versão do sistema

+++ server/src/utils/file_utils.go
@@ -2,20 +2,21 @@ package utils
 
 import (
 	"os"
-	"strings"
 	"path/filepath"
+	"strings"
 )
 
 const (
-	ImagesRoute 	= "http://localhost:8080/image"
-	ImagesDir 		= "../../public/"
+	imagesDirKey = "IMAGES_FILE_PATH"
 )
 
-func UploadFile(file []byte, filename string) (string, error) {
-	var imageFilepath = ImagesRoute + "/" + filename
-	absPath, _ := filepath.Abs(ImagesDir + filename)
+func UploadFile(file []byte, folderName, filename string) (string, error) {
+	var imageFilepath = GetImageRoute(folderName, filename)
+	var imagesDir = GetImagesFilePath()
 
-	err := os.WriteFile(absPath, file, 0644)
+	os.Mkdir(imagesDir+"/"+folderName, 0755)
+
+	err := os.WriteFile(imagesDir+folderName+"/"+filename, file, 0755)
 	if err != nil {
 		return "", err
 	}
@@ -23,11 +24,40 @@ func UploadFile(file []byte, filename string) (string, error) {
 	return imageFilepath, nil
 }
 
-func RemoveFile(onlinePath string) error {
-	filename := strings.Replace(onlinePath, ImagesRoute + "/", "", -1)
-	absPath, _ := filepath.Abs(ImagesDir + filename)
+func RemoveFile(filename, folderName string) error {
+	if !strings.Contains(filename, folderName) {
+		return nil
+	}
+
+	var imagesDir = GetImagesFilePath()
+
+	absPath, _ := filepath.Abs(imagesDir + folderName + "/" + filename)
 
 	err := os.Remove(absPath)
 
 	return err
+}
+
+func RemoveFolder(folder string) error {
+	var imagesDir = GetImagesFilePath()
+
+	absPath, _ := filepath.Abs(imagesDir + folder)
+
+	err := os.RemoveAll(absPath)
+
+	return err
+}
+
+func GetImagesFilePath() string {
+	var imagesDir = "../../public/" // for tests
+
+	if value, ok := os.LookupEnv(imagesDirKey); ok { // in container
+		imagesDir = value
+	}
+
+	return imagesDir
+}
+
+func GetImageRoute(folderName, filename string) string {
+	return "image/" + folderName + "/" + filename
 }
